Tests/LibWeb/Text/expected/XML/xml-utf32-encoding-ignored.txt

Normal XML: characterSet=UTF-8, element=root
UTF-32 declared: characterSet=UTF-8, element=root
UTF-32BE declared: characterSet=UTF-8, element=root
UTF-32LE declared: characterSet=UTF-8, element=root
